// src/building-profile/types/person-type.ts
//Tipo referente as pessoas
export type Person = {
  cpf?: string;
  rg?: string;
  name?: string;
  surname?: string;
  telphone?: string;
  cep?: string;
  state?: string;
  city?: string;
  district?: string;
  road?: string;
  number?: string | number;
  qrCode?: string;
  dateRegistration?: string;
};